requirements.txt
# Python Dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.18.4
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0

# Logging and utilities
colorlog==6.8.0
python-dateutil==2.8.2
pytz==2023.3
